+++ backend/internal/domain/donor/donor_repo.go
@@ -20,5 +20,5 @@ type Repository interface {
 	ExistsByDonorID(ctx context.Context, donorID string) (bool, error)
 	Count(ctx context.Context) (int, error)
 	GetByPetID(ctx context.Context, petID string) (*donormodel.DonorResponse, error)
-	Confirm(ctx context.Context, id string, factAmount int32) error
+	Confirm(ctx context.Context, id string, factAmount float64) error
 }
