+++ backend/internal/infra/ent/schema/donor_response.go
@@ -15,7 +15,7 @@ type DonorResponse struct {
 // Fields of the BloodSearchRequest.
 func (DonorResponse) Fields() []ent.Field {
 	return []ent.Field{
-		field.Int32("amount").Optional(),
+		field.Float("amount").Optional(),
 		field.Enum("compensation_type").Values("free", "paid", "food").Optional(),
 		field.Bool("taxi_compensation").Optional(),
 		field.Enum("status").Values("pending", "accepted", "rejected", "cancelled", "completed", "failed"),
