+++ backend/internal/infra/presistance/pg/donor_repo_ent.go
@@ -201,7 +201,7 @@ func (r *EntDonorResponseRepository) Count(ctx context.Context) (int, error) {
 }
 
 // Подтверждение донации реципиентом
-func (r *EntDonorResponseRepository) Confirm(ctx context.Context, donorResponseID string, factAmount int32) error {
+func (r *EntDonorResponseRepository) Confirm(ctx context.Context, donorResponseID string, factAmount float64) error {
 	update := r.client(ctx).DonorResponse.
 		UpdateOneID(donorResponseID).
 		SetIsConfirmed(true)
