+++ backend/internal/transport/http/dto/pet_dto.go
@@ -185,7 +185,7 @@ type PetDetail struct {
 	BreedID              string             `json:"breedId,omitempty" doc:"ID породы" example:"MIX"`
 	BloodGroup           string             `json:"bloodGroup,omitempty" doc:"Группа крови" enum:"DEA 1+,DEA 1-,A,B,AB" example:"DEA 1+"`
 	PetStatus            string             `json:"petStatus" doc:"Статус питомца" enum:"none,donor,recipient,blood_found,recovering,planned_donation" example:"donor"`
-	AvailableBloodAmount int32              `json:"availableBloodAmount,omitempty" doc:"Доступный объем крови для донации в мл" example:"450"`
+	AvailableBloodAmount float64            `json:"availableBloodAmount,omitempty" doc:"Доступный объем крови для донации в мл" example:"450"`
 	DonorRestrictions    *DonorRestrictions `json:"donorRestrictions,omitempty" doc:"Стоп-факторы и предупреждения"`
 	Health               *PetHealth         `json:"health,omitempty" doc:"Информация о здоровье"`
 	Treatments           *PetTreatment      `json:"treatments,omitempty" doc:"Информация о лечении"`
